[#32] refactor: 뒤로가기 버튼이 있는 헤더 컴포넌트 분리
