+++ src/pages/LoginPassword.tsx
@@ -1,10 +1,10 @@
-import { ReactComponent as BackArrow } from '../assets/image/icon/backArrow.svg';
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import { useAccountDispatch, useAccountState } from '../context/AccountContext.tsx';
 import axios from '../api/apiController.tsx';
 import Input from '../components/account/Input.tsx';
 import NextButton from '../components/account/NextButton.tsx';
+import HeaderWithBack from '../components/common/HeaderWithBack.tsx';
 
 function LoginPasswordInputPage() {
   const { email, password } = useAccountState();
@@ -40,17 +40,9 @@ function LoginPasswordInputPage() {
       });
   };
 
-  const onClickGoBack = () => {
-    navigate(-1);
-  };
-
   return (
     <>
-      <Header>
-        <BackButton onClick={onClickGoBack}>
-          <BackArrow />
-        </BackButton>
-      </Header>
+      <HeaderWithBack />
       <Head1>비밀번호를 입력해주세요.</Head1>
       <Input type="password" value={password} onChange={onInputPassword} />
       <Body2>영문, 숫자, 특수문자를 포함해 8자 이상</Body2>
@@ -73,20 +65,6 @@ const Head1 = styled.div`
   color: #ffffff;
   font-family: 'Pretendard';
   margin-bottom: 17px;
-`;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 11.5px 0;
-  height: 44px;
-  margin-bottom: 24px;
-`;
-
-const BackButton = styled.button`
-  position: absolute;
-  left: 16px;
 `;
 
 const Body2 = styled.div`
